local version 2

+++ backend/foodgram/api_recipes/serializers.py
@@ -1,18 +1,19 @@
 from drf_extra_fields import fields as drf_fields
 from rest_framework import serializers
 
-from api_user.serializers import UserAccountSerializer
+from api_user.serializers import CustomUserSerializer
 from core.models import (
-    Favorite,
+    UserFavoriteRecipe,
     Ingredient,
-    IngredientInRecipe,
-    Recipe,
-    ShoppingCart,
+    RecipeIngredient,
+    CulinaryRecipe,
+    UserShoppingCart,
     User
 )
 
 
 class IngredientSerializer(serializers.ModelSerializer):
+    """Сериализатор для ингредиентов."""
     class Meta:
         model = Ingredient
         fields = [
@@ -23,6 +24,7 @@ class IngredientSerializer(serializers.ModelSerializer):
 
 
 class IngredientInRecipeSerializer(serializers.ModelSerializer):
+    """Сериализатор для отображения ингредиентов в рецепте."""
     id = serializers.ReadOnlyField(
         source='ingredient.id'
     )
@@ -34,7 +36,7 @@ class IngredientInRecipeSerializer(serializers.ModelSerializer):
     )
 
     class Meta:
-        model = IngredientInRecipe
+        model = RecipeIngredient
         fields = [
             'id',
             'name',
@@ -53,15 +55,15 @@ class CreateIngredientInRecipeSerializer(serializers.ModelSerializer):
     )
 
     class Meta:
-        model = IngredientInRecipe
+        model = RecipeIngredient
         fields = [
             'id',
             'amount'
         ]
 
 
-class RecipeSerializer(serializers.ModelSerializer):
-    author = UserAccountSerializer()
+class RecipeDetailSerializer(serializers.ModelSerializer):
+    author = CustomUserSerializer()
     ingredients = IngredientInRecipeSerializer(
         source='ingredient_amounts',
         many=True
@@ -76,7 +78,7 @@ class RecipeSerializer(serializers.ModelSerializer):
     def get_is_favorited(self, obj):
         user = self.context.get('request').user
         if user:
-            return user.is_authenticated and Favorite.objects.filter(
+            return user.is_authenticated and UserFavoriteRecipe.objects.filter(
                 user__exact=user,
                 recipe__exact=obj
             ).exists()
@@ -85,14 +87,14 @@ class RecipeSerializer(serializers.ModelSerializer):
     def get_is_in_shopping_cart(self, obj):
         user = self.context.get('request').user
         if user:
-            return user.is_authenticated and ShoppingCart.objects.filter(
+            return user.is_authenticated and UserShoppingCart.objects.filter(
                 user__exact=user,
                 recipe__exact=obj
             ).exists()
         return False
 
     class Meta:
-        model = Recipe
+        model = CulinaryRecipe
         fields = [
             'id',
             'author',
@@ -106,14 +108,14 @@ class RecipeSerializer(serializers.ModelSerializer):
         ]
 
 
-class CreateRecipeSerializer(serializers.ModelSerializer):
+class RecipeCreateUpdateSerializer(serializers.ModelSerializer):
     ingredients = CreateIngredientInRecipeSerializer(
         many=True
     )
     image = drf_fields.Base64ImageField()
 
     class Meta:
-        model = Recipe
+        model = CulinaryRecipe
         fields = [
             'name',
             'image',
@@ -125,56 +127,58 @@ class CreateRecipeSerializer(serializers.ModelSerializer):
     def validate_image(self, image):
         if not image:
             raise serializers.ValidationError(
-                'Картинка является обязательным полем.'
+                'Изображение рецепта обязательно для заполнения.'
             )
         return image
 
     def validate_ingredients(self, ingredients):
         if not ingredients:
             raise serializers.ValidationError(
-                'Ингредиенты являются обязательным полем.'
+                'Необходимо добавить хотя бы один ингредиент.'
             )
-        ids = [ingredient['ingredient'].id for ingredient in ingredients]
-        if len(ids) != len(set(ids)):
+        ingredient_ids = [item['ingredient'].id for item in ingredients]
+        if len(ingredient_ids) != len(set(ingredient_ids)):
             raise serializers.ValidationError(
                 'Ингредиенты должны быть уникальны.'
             )
         return ingredients
 
-    def add_ingredients(self, recipe, ingredients):
-        return IngredientInRecipe.objects.bulk_create([
-            IngredientInRecipe(
+    def _create_recipe_ingredients(self, recipe, ingredients_data):
+        ingredients = [
+            RecipeIngredient(
                 recipe=recipe,
-                ingredient=ingredient['ingredient'],
-                amount=ingredient['amount']
+                ingredient=item['ingredient'],
+                amount=item['amount']
             )
-            for ingredient in ingredients
-        ])
+            for item in ingredients_data
+        ]
+        return RecipeIngredient.objects.bulk_create(ingredients)
 
     def create(self, validated_data):
-        ingredients = validated_data.pop('ingredients')
-        recipe = Recipe.objects.create(**validated_data)
-        self.add_ingredients(recipe, ingredients)
+        ingredients_data = validated_data.pop('ingredients')
+        recipe = CulinaryRecipe.objects.create(**validated_data)
+        self._create_recipe_ingredients(recipe, ingredients_data)
         return recipe
 
     def update(self, instance, validated_data):
-        ingredients = self.validate_ingredients(
+        ingredients_data = self.validate_ingredients(
             validated_data.pop('ingredients', None)
         )
-        instance.ingredient_amounts.all().delete()
-        self.add_ingredients(instance, ingredients)
+        if ingredients_data is not None:
+            instance.ingredient_amounts.all().delete()
+            self._create_recipe_ingredients(instance, ingredients_data)
         return super().update(instance, validated_data)
 
     def to_representation(self, instance):
-        return RecipeSerializer(
+        return RecipeDetailSerializer(
             instance,
             context=self.context
         ).data
 
 
-class ShortRecipeSerializer(serializers.ModelSerializer):
+class RecipeSummarySerializer(serializers.ModelSerializer):
     class Meta:
-        model = Recipe
+        model = CulinaryRecipe
         fields = [
             'id',
             'name',
@@ -183,7 +187,7 @@ class ShortRecipeSerializer(serializers.ModelSerializer):
         ]
 
 
-class UserWithRecipeSerializer(UserAccountSerializer):
+class UserSubscriptionSerializer(CustomUserSerializer):
     recipes = serializers.SerializerMethodField(
         method_name='get_recipes'
     )
@@ -212,7 +216,7 @@ class UserWithRecipeSerializer(UserAccountSerializer):
         recipes_limit = query_params.get('recipes_limit')
         if recipes_limit and recipes_limit.isdigit():
             recipes = recipes[:int(recipes_limit)]
-        return ShortRecipeSerializer(
+        return RecipeSummarySerializer(
             recipes,
             context={'request': request},
             many=True
@@ -221,7 +225,7 @@ class UserWithRecipeSerializer(UserAccountSerializer):
 
 class FavoriteSerializer(serializers.ModelSerializer):
     class Meta:
-        model = Favorite
+        model = UserFavoriteRecipe
         fields = [
             'user',
             'recipe'
@@ -230,7 +234,7 @@ class FavoriteSerializer(serializers.ModelSerializer):
 
 class ShoppingCartSerializer(serializers.ModelSerializer):
     class Meta:
-        model = ShoppingCart
+        model = UserShoppingCart
         fields = [
             'user',
             'recipe'
